backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["3R11K <erikbatista1222@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
flask = "^3.1.0"
pyjwt = "^2.10.0"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
